for nadav love jeremy

+++ neurodoll.py
@@ -417,8 +417,10 @@ def analyze_graylevels(url_or_np_array,labels=constants.ultimate_21):
 
     gl = get_all_category_graylevels(url_or_np_array)
     mask = gl.argmax(axis=2)
-    background = np.array((mask==0)*1)
-    foreground = np.array((mask>0)*1)
+    background = np.array((mask==0)*1,dtype=np.uint8)
+    foreground = np.array((mask>0)*1,dtype=np.uint8)
+#    astype(np.uint8))
+
     print('masktype: '+str(type(background))+' shape:'+str(foreground.shape))
     h,w = gl.shape[0:2]
     window_size = 1700
